tests: adjust tests to increase coverage
RandomStringClientTest:

Set baseUrl and len using reflection to ensure private fields are correctly initialized during tests.
Removed @SpringBootTest annotation to prevent application context conflicts and improve test isolation.
Corrected exception expectations to anticipate BadGatewayException instead of IllegalArgumentException.
Updated HttpClient mocks to accurately simulate IOException and InterruptedException scenarios.
StreamLambdaHandlerTest:

Refactored StreamLambdaHandler by extracting the static initializer into a protected method for enhanced testability without static mocking frameworks.
Added tests to cover exception handling in the static initializer, increasing line coverage related to initialization errors.
Implemented tests that simulate exceptions in the handleRequest method to verify proper error flow handling.
GlobalExceptionHandler:

Fixed the createErrorResponse method to utilize the isolatedErrorMessage instead of ex.getMessage(), ensuring accurate and user-friendly error messages.
Updated the test handleInvalidAttributeException_shouldReturnInvalidAttributeEntity to reflect the changes and verify that "Invalid value" is returned as expected.

+++ calculatorCreateOperation/src/main/java/org/exercise/core/dtos/Values.java
@@ -1,4 +1,4 @@
 package org.exercise.core.dtos;
 
-public record Values(Integer value1, Integer value2) {
+public record Values(Double value1, Double value2) {
 }
